Move file parsing out of wm module
